requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
ulid==1.1
PyJWT==2.8.0
cryptography==41.0.5
django==4.2.6
djangorestframework==3.14.0
django-extensions==3.2.3
pillow==10.1.0
requests==2.31.0
gunicorn==20.1.0
django-tinymce==3.6.1
pandas==1.4.0
django-ckeditor==6.7.0
mysqlclient==2.1.0
django-multiselectfield==0.1.12
beautifulsoup4==4.12.2
markdown==3.5.1
jinja2==3.1.2